# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: A reselling marketplace backend with authentication, product management, cart, and checkout functionality
  version: 1.0.0
  contact:
    name: Marketplace API Support
    email: support@marketplace.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-deployment-url.com
    description: Production server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        condition:
          type: string
          enum: [New, LikeNew, Good, Fair, Poor, Used]
        availability:
          type: boolean
        quantity:
          type: integer
        sellerId:
          type: integer
        categoryId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalCount:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
security:
  - cookieAuth: []
paths:
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: Get all products with pagination and filters
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: category
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: condition
          in: query
          schema:
            type: string
            enum: [New, LikeNew, Good, Fair, Poor, Used]
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      tags: [Products]
      summary: Create a new product (Seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, categoryId]
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: float
                condition:
                  type: string
                  enum: [New, LikeNew, Good, Fair, Poor, Used]
                quantity:
                  type: integer
                  minimum: 1
                categoryId:
                  type: integer
                images:
                  type: array
                  items:
                    type: string
                    format: base64
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  product:
                    $ref: '#/components/schemas/Product'
...